requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
schedule==1.2.2
APScheduler==3.10.4
gunicorn==21.2.0
discord.py==2.6.3
twooter
openai==1.109.1
anthropic==0.68.0
google-generativeai==0.8.5
